refactor(comments): Stage 3 — extract ReplyComposer shared component
Consolidates the composer UI (contenteditable + @mentions + media upload
+ GIF picker + poll creator + submit/cancel) previously duplicated across
four components into a single src/components/comments/ReplyComposer.svelte.

Migration order (containers first, then inline composers):
  1. FeedComments.svelte  → ReplyComposer (implicit + silent + compact)
  2. Comments.svelte      → ReplyComposer (explicit + alert + optimistic add)
  3. FeedComment.svelte   → ReplyComposer (+ replyTo, showCancel, rootEvent prop)
  4. Comment.svelte       → ReplyComposer (+ replyTo, showCancel, rootEvent prop)

Root-event plumbing: Comment.svelte and FeedComment.svelte gain a new
`rootEvent: NDKEvent` prop passed down from their containers (Comments /
FeedComments). This lets the inline composers hand the root (recipe / feed
post) to postComment() as parentEvent, with the comment-being-replied-to
as replyTo — matching the spec-compliant Stage 2 API shape. Previously,
the inline composers reconstructed a synthetic root-event object from the
parent comment's A/a tags inline; that ~40-line reconstruction block is
deleted along with the rest of the inlined composer logic.

Behavior preserved per caller (Stage 5 unifies):
  - Comments: optimistic add via onPosted callback; alert() on error; <Button>
    component for submit via slot.
  - FeedComments: subscription round-trip only; silent error handling; compact;
    plain .btn-post submit.
  - Comment inline reply: alert() on error (matches Comments.svelte's
    stated recipe-context UX); compact + showCancel.
  - FeedComment inline reply: silent error; compact + showCancel.

Submit-button slot: `<slot name="submit" submit={handleSubmit} disabled
={isDisabled} posting />` lets callers that want the <Button> component
(Comments.svelte) override the default plain .btn-post markup without
coupling ReplyComposer to any particular button implementation.

Stage-0 bug fixes landed as side effects of extraction:
  - Comment.svelte's postingReply-stuck-on-error bug (scheduled for
    Stage 4) is eliminated — ReplyComposer's internal `posting` flag is
    managed via try/finally in the new code, which correctly resets on
    error. Stage 4 scope updated to remove this fix from its list.
  - Comment.svelte previously errored silently (no alert, no state reset,
    button frozen). Stage 3 routes errors through onErrorStrategy='alert'
    for consistency with the Comments container's stated intent. This is
    a minor user-visible improvement the refactor exposed.

Preserved: refresh() callback prop on inline components (now a no-op
passed from container, kept as a stable API for the GifPicker/PollCreator
events); Stage 4 deletes it when merging Comment + FeedComment → CommentCard.

Spec compliance: postComment unchanged. All four callers pass
parentEvent (root) + optional replyTo exactly as Stage 2 specified, so
NIP-22 kind-1111 on kind-30023 recipes and NIP-10 kind-1 on kind-1 feed
posts are byte-for-byte equivalent to Stage 2 output.

Verified after each migration step:
  - pnpm run check: 4 pre-existing errors / 127 warnings (2 fewer than
    pre-Stage-3; deleted .ring:/.ring-color CSS in FeedComment).
  - pnpm run build: passes.
  - Dev server: /, /polls, /reads all 200 with no runtime errors.

Line-count delta:
  - Comment.svelte:      886 → 457 (−429)
  - FeedComment.svelte:  893 → 470 (−423)
  - Comments.svelte:     387 →  99 (−288)
  - FeedComments.svelte: 390 → 113 (−277)
  - New ReplyComposer.svelte:      547
  - Net: −870 lines (Stage 0 estimated −500).

Cumulative Task 6 progress: Stage 1 (−1,293) + Stage 2 (+311) + Stage 3
(−870) = −1,852 lines. Stage 4 (CommentCard merge) and Stage 5
(CommentThread merge + toast UX) still ahead.
