+++ src/components/Comment.svelte
@@ -5,30 +5,24 @@
   import { format as formatDate } from 'timeago.js';
   import Avatar from './Avatar.svelte';
   import ZapModal from './ZapModal.svelte';
+  import ReplyComposer from './comments/ReplyComposer.svelte';
   import { resolveProfileByPubkey, formatDisplayName } from '$lib/profileResolver';
   import { formatAmount } from '$lib/utils';
   import { addClientTagToEvent } from '$lib/nip89';
-  import { buildNip22CommentTags } from '$lib/tagUtils';
   import HeartIcon from 'phosphor-svelte/lib/Heart';
   import LightningIcon from 'phosphor-svelte/lib/Lightning';
   import NoteContent from './NoteContent.svelte';
   import { onMount, onDestroy } from 'svelte';
-  import { get } from 'svelte/store';
   import { extractZapAmountSats } from '$lib/zapAmount';
-  import MentionDropdown from './MentionDropdown.svelte';
-  import { MentionComposerController, type MentionState } from '$lib/mentionComposer';
-  import GifIcon from 'phosphor-svelte/lib/Gif';
-  import ImageIcon from 'phosphor-svelte/lib/Image';
-  import VideoIcon from 'phosphor-svelte/lib/Video';
-  import GifPicker from './GifPicker.svelte';
-  import ChartBarHorizontalIcon from 'phosphor-svelte/lib/ChartBarHorizontal';
-  import PollCreator from './PollCreator.svelte';
-  import { buildPollTags, type PollConfig } from '$lib/polls';
-  import { uploadImage, uploadVideo } from '$lib/mediaUpload';
 
   export let event: NDKEvent;
   export let allReplies: NDKEvent[] = []; // All replies for finding parent
   export let refresh: () => void;
+  /**
+   * The root event (recipe) passed down from the Comments container. Used by
+   * the inline reply composer to build NIP-22 tags with the correct root-scope.
+   */
+  export let rootEvent: NDKEvent;
 
   // Display state
   let displayName = '';
@@ -39,48 +33,8 @@
   let parentDisplayName = '';
   let parentLoading = true;
 
-  // Reply box state
+  // Reply box state — the composer itself owns everything inside the form.
   let showReplyBox = false;
-  let showGifPicker = false;
-  let showPollCreator = false;
-  let pollConfig: PollConfig | null = null;
-  let replyText = '';
-  let postingReply = false;
-  let replyComposerEl: HTMLDivElement;
-  let lastRenderedReply = '';
-  let uploadedImages: string[] = [];
-  let uploadedVideos: string[] = [];
-  let uploadingImage = false;
-  let uploadingVideo = false;
-  let uploadError = '';
-  let imageInputEl: HTMLInputElement;
-  let videoInputEl: HTMLInputElement;
-
-  // Mention autocomplete
-  let mentionState: MentionState = {
-    mentionQuery: '',
-    showMentionSuggestions: false,
-    mentionSuggestions: [],
-    selectedMentionIndex: 0,
-    mentionSearching: false
-  };
-
-  const mentionCtrl = new MentionComposerController(
-    (state) => {
-      mentionState = state;
-    },
-    (text) => {
-      replyText = text;
-      lastRenderedReply = text;
-    }
-  );
-
-  $: mentionCtrl.setComposerEl(replyComposerEl);
-
-  $: if (replyComposerEl && replyText !== lastRenderedReply) {
-    mentionCtrl.syncContent(replyText);
-    lastRenderedReply = replyText;
-  }
 
   // Like state
   let liked = false;
@@ -103,8 +57,6 @@
     const replyTag = eTags.find((tag) => tag[3] === 'reply');
     if (replyTag) {
       // Check if this reply tag points to another comment (not the root recipe)
-      const aTag = event.getMatchingTags('a')[0];
-      // If it's a reply and there's another e tag, it might be a nested reply
       const parentEventTag = eTags.find((tag) => tag[3] !== 'reply' && tag[3] !== 'root');
       if (parentEventTag) return parentEventTag[1];
 
@@ -118,11 +70,6 @@
 
   // Load profile and parent comment
   onMount(async () => {
-    // Preload mention profiles in background
-    if ($userPublickey) {
-      mentionCtrl.preloadFollowList();
-    }
-
     // Load author profile
     if (event.pubkey && $ndk) {
       try {
@@ -238,182 +185,12 @@
     likeCount++;
   }
 
-  async function handleImageUpload(e: Event) {
-    const target = e.target as HTMLInputElement;
-    const files = target.files;
-    if (!files || files.length === 0) return;
-    uploadingImage = true;
-    uploadError = '';
-    try {
-      for (const file of Array.from(files)) {
-        const url = await uploadImage($ndk, file);
-        uploadedImages = [...uploadedImages, url];
-      }
-    } catch (err: any) {
-      uploadError = err?.message || 'Failed to upload image.';
-    } finally {
-      uploadingImage = false;
-      if (imageInputEl) imageInputEl.value = '';
-    }
-  }
-
-  async function handleVideoUpload(e: Event) {
-    const target = e.target as HTMLInputElement;
-    const files = target.files;
-    if (!files || files.length === 0) return;
-    uploadingVideo = true;
-    uploadError = '';
-    try {
-      for (const file of Array.from(files)) {
-        const url = await uploadVideo($ndk, file);
-        uploadedVideos = [...uploadedVideos, url];
-      }
-    } catch (err: any) {
-      uploadError = err?.message || 'Failed to upload video.';
-    } finally {
-      uploadingVideo = false;
-      if (videoInputEl) videoInputEl.value = '';
-    }
-  }
-
-  function removeImage(index: number) {
-    uploadedImages = uploadedImages.filter((_, i) => i !== index);
-  }
-
-  function removeVideo(index: number) {
-    uploadedVideos = uploadedVideos.filter((_, i) => i !== index);
-  }
-
-  // Post reply
-  async function postReply() {
-    if ((!replyText.trim() && uploadedImages.length === 0 && uploadedVideos.length === 0 && !pollConfig) || postingReply) return;
-
-    postingReply = true;
-    const ev = new NDKEvent($ndk);
-
-    // Check if this is a reply to a recipe comment
-    // If the parent comment is kind 1111 or has an 'a' tag referencing kind 30023, use kind 1111
-    const aTag = event.getMatchingTags('a')[0];
-    const ATag = event.getMatchingTags('A')[0];
-    const isRecipeReply =
-      event.kind === 1111 ||
-      (aTag && aTag[1]?.startsWith('30023:')) ||
-      (ATag && ATag[1]?.startsWith('30023:'));
-    ev.kind = pollConfig ? 1068 : (isRecipeReply ? 1111 : 1);
-
-    if (replyComposerEl) {
-      replyText = mentionCtrl.extractText();
-      lastRenderedReply = replyText;
-    }
-
-    let replyContent = mentionCtrl.replacePlainMentions(replyText.trim());
-    const mediaUrls = [...uploadedImages, ...uploadedVideos];
-    if (mediaUrls.length > 0) {
-      const mediaText = mediaUrls.join('\n');
-      replyContent = replyContent ? `${replyContent}\n\n${mediaText}` : mediaText;
-    }
-    ev.content = replyContent;
-
-    // Reconstruct a minimal event object for the parent comment
-    const parentEventObj = {
-      id: event.id,
-      pubkey: event.pubkey,
-      kind: event.kind,
-      tags: event.tags
-    };
-
-    // For recipe replies, we need to get the root event info from the parent's tags
-    if (isRecipeReply) {
-      // Get root event info from parent comment's tags
-      const rootATag = event.getMatchingTags('A')[0] || event.getMatchingTags('a')[0];
-
-      if (rootATag) {
-        // Parse the address tag to extract root event info
-        const [kind, pubkey, ...dTagParts] = rootATag[1].split(':');
-        const dTag = dTagParts.join(':');
-        const rootEventObj = {
-          kind: parseInt(kind),
-          pubkey: pubkey,
-          id: '', // We don't need the actual event ID for tag generation
-          tags: [
-            ['d', dTag],
-            ['relay', rootATag[2] || '']
-          ]
-        };
-
-        // Use the utility with both root and parent event
-        ev.tags = buildNip22CommentTags(rootEventObj, {
-          ...parentEventObj,
-          tags: parentEventObj.tags as string[][]
-        });
-      } else {
-        // Fallback: treat parent as if it's the root
-        ev.tags = buildNip22CommentTags(
-          {
-            ...parentEventObj,
-            kind: parentEventObj.kind ?? 1,
-            tags: parentEventObj.tags as string[][]
-          },
-          {
-            ...parentEventObj,
-            tags: parentEventObj.tags as string[][]
-          }
-        );
-      }
-    } else {
-      // For non-recipe replies, we still need to construct a root event object
-      // In this case, we need to find the root event ID from the parent's tags
-      const rootETag = event.getMatchingTags('e').find((t) => t[3] === 'root');
-      if (rootETag) {
-        // Derive the root author's pubkey from the parent's p-tags when possible
-        const rootPTags = event.getMatchingTags('p');
-        const rootPubkey = rootPTags && rootPTags.length > 0 ? rootPTags[0][1] : event.pubkey;
-        const rootEventObj = {
-          kind: 1,
-          pubkey: rootPubkey, // Use root author's pubkey when available, otherwise fall back to parent
-          id: rootETag[1], // Root event ID from the e tag
-          tags: []
-        };
-        ev.tags = buildNip22CommentTags(rootEventObj, parentEventObj);
-      } else {
-        // Use the simplified version for replies
-        ev.tags = [
-          ['e', event.id, '', 'reply'],
-          ['p', event.pubkey]
-        ];
-      }
-    }
-
-    // Parse and add @ mention tags (p tags)
-    const mentions = mentionCtrl.parseMentions(replyContent);
-    for (const pubkey of mentions.values()) {
-      if (!ev.tags.some((t) => t[0] === 'p' && t[1] === pubkey)) {
-        ev.tags.push(['p', pubkey]);
-      }
-    }
-
-    addClientTagToEvent(ev);
-    if (pollConfig) {
-      ev.tags.push(...buildPollTags(pollConfig));
-    }
-    await ev.publish();
-    replyText = '';
-    lastRenderedReply = '';
-    uploadedImages = [];
-    uploadedVideos = [];
-    pollConfig = null;
-    uploadError = '';
-    if (replyComposerEl) {
-      replyComposerEl.innerHTML = '';
-    }
-    mentionCtrl.resetMentionState();
+  function handleReplyPosted() {
     showReplyBox = false;
-    postingReply = false;
     refresh();
   }
 
   onDestroy(() => {
-    mentionCtrl.destroy();
     if (likeSubscription) {
       likeSubscription.stop();
     }
@@ -521,127 +298,19 @@
         </button>
       </div>
 
-      <!-- Inline Reply Box -->
+      <!-- Inline Reply Composer -->
       {#if showReplyBox}
-        <div class="reply-form">
-          <div class="relative">
-            <div
-              bind:this={replyComposerEl}
-              class="reply-input"
-              contenteditable={!postingReply}
-              role="textbox"
-              tabindex="0"
-              aria-multiline="true"
-              data-placeholder="Add a reply..."
-              on:input={() => mentionCtrl.handleInput()}
-              on:keydown={(e) => mentionCtrl.handleKeydown(e)}
-              on:beforeinput={(e) => mentionCtrl.handleBeforeInput(e)}
-              on:paste={(e) => mentionCtrl.handlePaste(e)}
-            ></div>
-
-            <MentionDropdown
-              show={mentionState.showMentionSuggestions}
-              suggestions={mentionState.mentionSuggestions}
-              selectedIndex={mentionState.selectedMentionIndex}
-              searching={mentionState.mentionSearching}
-              query={mentionState.mentionQuery}
-              on:select={(e) => mentionCtrl.insertMention(e.detail)}
-            />
-          </div>
-          {#if uploadError}
-            <p class="text-red-500 text-xs">{uploadError}</p>
-          {/if}
-
-          {#if uploadedImages.length > 0}
-            <div class="flex flex-wrap gap-2">
-              {#each uploadedImages as imageUrl, index}
-                <div class="relative group">
-                  <img src={imageUrl} alt="Upload preview" class="w-16 h-16 object-cover rounded-lg" style="border: 1px solid var(--color-input-border)" />
-                  <button type="button" on:click={() => removeImage(index)} class="absolute -top-2 -right-2 bg-red-500 hover:bg-red-600 text-white rounded-full p-0.5 shadow-lg" aria-label="Remove image">
-                    <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" /></svg>
-                  </button>
-                </div>
-              {/each}
-            </div>
-          {/if}
-
-          {#if uploadedVideos.length > 0}
-            <div class="flex flex-wrap gap-2">
-              {#each uploadedVideos as videoUrl, index}
-                <div class="relative group">
-                  <video src={videoUrl} class="w-24 h-16 object-cover rounded-lg" style="border: 1px solid var(--color-input-border)" preload="metadata" muted />
-                  <button type="button" on:click={() => removeVideo(index)} class="absolute -top-2 -right-2 bg-red-500 hover:bg-red-600 text-white rounded-full p-0.5 shadow-lg" aria-label="Remove video">
-                    <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" /></svg>
-                  </button>
-                </div>
-              {/each}
-            </div>
-          {/if}
-
-          <div class="reply-buttons">
-            <label class="btn-media" class:opacity-50={uploadingImage || uploadingVideo || postingReply} title="Upload image">
-              <ImageIcon size={16} />
-              <input bind:this={imageInputEl} type="file" accept="image/*" class="sr-only" on:change={handleImageUpload} disabled={postingReply || uploadingImage || uploadingVideo} />
-            </label>
-            <label class="btn-media" class:opacity-50={uploadingImage || uploadingVideo || postingReply} title="Upload video">
-              <VideoIcon size={16} />
-              <input bind:this={videoInputEl} type="file" accept="video/*" class="sr-only" on:change={handleVideoUpload} disabled={postingReply || uploadingImage || uploadingVideo} />
-            </label>
-            <button
-              on:click={() => (showGifPicker = true)}
-              class="btn-gif"
-              title="Add GIF"
-              disabled={postingReply || uploadingImage || uploadingVideo}
-            >
-              <GifIcon size={16} />
-            </button>
-            <button
-              on:click={() => (showPollCreator = true)}
-              class="btn-gif"
-              title="Create poll"
-              disabled={postingReply}
-              class:opacity-50={postingReply}
-            >
-              <ChartBarHorizontalIcon size={16} class={pollConfig ? 'text-primary' : ''} />
-            </button>
-            {#if uploadingImage}
-              <span class="text-xs text-caption">Uploading image...</span>
-            {:else if uploadingVideo}
-              <span class="text-xs text-caption">Uploading video...</span>
-            {/if}
-            {#if pollConfig}
-              <span class="text-xs text-orange-600 flex items-center gap-1">
-                <ChartBarHorizontalIcon size={12} />
-                Poll ({pollConfig.options.length})
-                <button type="button" on:click={() => (pollConfig = null)} class="hover:text-orange-800">×</button>
-              </span>
-            {/if}
-            <button
-              on:click={postReply}
-              disabled={(!replyText.trim() && uploadedImages.length === 0 && uploadedVideos.length === 0 && !pollConfig) || postingReply || uploadingImage || uploadingVideo}
-              class="btn-post"
-            >
-              {postingReply ? 'Posting...' : 'Post'}
-            </button>
-            <button
-              on:click={() => {
-                showReplyBox = false;
-                replyText = '';
-                lastRenderedReply = '';
-                uploadedImages = [];
-                uploadedVideos = [];
-                uploadError = '';
-                if (replyComposerEl) {
-                  replyComposerEl.innerHTML = '';
-                }
-                mentionCtrl.resetMentionState();
-              }}
-              class="btn-cancel"
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
+        <ReplyComposer
+          parentEvent={rootEvent}
+          replyTo={event}
+          placeholder="Add a reply..."
+          signingStrategy="explicit-with-timeout"
+          onErrorStrategy="alert"
+          compact
+          showCancel
+          onPosted={handleReplyPosted}
+          on:cancel={() => (showReplyBox = false)}
+        />
       {/if}
     </div>
   </div>
@@ -651,20 +320,6 @@
 {#if zapModalOpen}
   <ZapModal bind:open={zapModalOpen} {event} />
 {/if}
-
-<GifPicker
-  bind:open={showGifPicker}
-  on:select={(e) => {
-    uploadedImages = [...uploadedImages, e.detail.url];
-  }}
-/>
-
-<PollCreator
-  bind:open={showPollCreator}
-  on:create={(e) => {
-    pollConfig = e.detail;
-  }}
-/>
 
 <style>
   /* Comment card - full width, no nesting */
@@ -798,89 +453,5 @@
     align-items: center;
     gap: 0.25rem;
     color: var(--color-text-secondary);
-  }
-
-  /* Reply form */
-  .reply-form {
-    margin-top: 0.75rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .reply-input {
-    width: 100%;
-    padding: 0.5rem 0.75rem;
-    font-size: 1rem;
-    border-radius: 0.5rem;
-    background: var(--color-input-bg);
-    border: 1px solid var(--color-input-border);
-    color: var(--color-text-primary);
-    white-space: pre-wrap;
-    word-break: break-word;
-  }
-
-  .reply-input:focus {
-    outline: none;
-    box-shadow: 0 0 0 2px var(--color-primary);
-  }
-
-  .reply-input:empty:before {
-    content: attr(data-placeholder);
-    color: var(--color-caption);
-    pointer-events: none;
-  }
-
-  .reply-buttons {
-    display: flex;
-    gap: 0.5rem;
-  }
-
-  .btn-post {
-    padding: 0.5rem 1rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-    color: white;
-    background: var(--color-primary);
-    border-radius: 0.5rem;
-  }
-
-  .btn-post:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-  }
-
-  .btn-cancel {
-    padding: 0.5rem 1rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-    color: var(--color-text-primary);
-    background: var(--color-input-bg);
-    border-radius: 0.5rem;
-  }
-
-  .btn-cancel:hover {
-    opacity: 0.8;
-  }
-
-  .btn-gif,
-  .btn-media {
-    padding: 0.375rem;
-    color: var(--color-caption);
-    border-radius: 0.375rem;
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-    transition: opacity 0.15s;
-  }
-
-  .btn-gif:hover,
-  .btn-media:hover {
-    opacity: 0.7;
-  }
-
-  .btn-gif:disabled {
-    opacity: 0.4;
-    cursor: not-allowed;
   }
 </style>
