+++ src/components/FeedComment.svelte
@@ -7,30 +7,25 @@
   import CustomAvatar from './CustomAvatar.svelte';
   import NoteContent from './NoteContent.svelte';
   import ZapModal from './ZapModal.svelte';
+  import ReplyComposer from './comments/ReplyComposer.svelte';
   import { resolveProfileByPubkey, formatDisplayName } from '$lib/profileResolver';
   import { formatAmount } from '$lib/utils';
   import HeartIcon from 'phosphor-svelte/lib/Heart';
   import LightningIcon from 'phosphor-svelte/lib/Lightning';
   import { addClientTagToEvent } from '$lib/nip89';
-  import { buildNip22CommentTags } from '$lib/tagUtils';
   import { onMount, onDestroy } from 'svelte';
-  import { get } from 'svelte/store';
   import { extractZapAmountSats } from '$lib/zapAmount';
-  import MentionDropdown from './MentionDropdown.svelte';
-  import { MentionComposerController, type MentionState } from '$lib/mentionComposer';
-  import GifIcon from 'phosphor-svelte/lib/Gif';
-  import ImageIcon from 'phosphor-svelte/lib/Image';
-  import VideoIcon from 'phosphor-svelte/lib/Video';
-  import GifPicker from './GifPicker.svelte';
-  import ChartBarHorizontalIcon from 'phosphor-svelte/lib/ChartBarHorizontal';
-  import PollCreator from './PollCreator.svelte';
-  import { buildPollTags, type PollConfig } from '$lib/polls';
-  import { uploadImage, uploadVideo } from '$lib/mediaUpload';
 
   export let event: NDKEvent;
   export let allComments: NDKEvent[] = []; // All comments for finding parent
   export let refresh: () => void;
   export let mainEventId: string;
+  /**
+   * The root event (feed post) passed down from the FeedComments container.
+   * Used by the inline reply composer to build NIP-22/NIP-10 tags with the
+   * correct root-scope.
+   */
+  export let rootEvent: NDKEvent;
 
   // Display state
   let displayName = '';
@@ -41,48 +36,8 @@
   let parentDisplayName = '';
   let parentLoading = true;
 
-  // Reply box state
+  // Reply box state — the composer itself owns everything inside the form.
   let showReplyBox = false;
-  let showGifPicker = false;
-  let showPollCreator = false;
-  let pollConfig: PollConfig | null = null;
-  let replyText = '';
-  let postingReply = false;
-  let replyComposerEl: HTMLDivElement;
-  let lastRenderedReply = '';
-  let uploadedImages: string[] = [];
-  let uploadedVideos: string[] = [];
-  let uploadingImage = false;
-  let uploadingVideo = false;
-  let uploadError = '';
-  let imageInputEl: HTMLInputElement;
-  let videoInputEl: HTMLInputElement;
-
-  // Mention autocomplete
-  let mentionState: MentionState = {
-    mentionQuery: '',
-    showMentionSuggestions: false,
-    mentionSuggestions: [],
-    selectedMentionIndex: 0,
-    mentionSearching: false
-  };
-
-  const mentionCtrl = new MentionComposerController(
-    (state) => {
-      mentionState = state;
-    },
-    (text) => {
-      replyText = text;
-      lastRenderedReply = text;
-    }
-  );
-
-  $: mentionCtrl.setComposerEl(replyComposerEl);
-
-  $: if (replyComposerEl && replyText !== lastRenderedReply) {
-    mentionCtrl.syncContent(replyText);
-    lastRenderedReply = replyText;
-  }
 
   // Like state
   let liked = false;
@@ -145,9 +100,6 @@
 
   // Load profile and parent comment
   onMount(async () => {
-    // Preload follow list for mention autocomplete
-    mentionCtrl.preloadFollowList();
-
     // Load author profile
     if (event.pubkey && $ndk) {
       try {
@@ -219,7 +171,6 @@
   });
 
   onDestroy(() => {
-    mentionCtrl.destroy();
     if (likeSubscription) {
       likeSubscription.stop();
     }
@@ -249,171 +200,9 @@
     }
   }
 
-  async function handleImageUpload(e: Event) {
-    const target = e.target as HTMLInputElement;
-    const files = target.files;
-    if (!files || files.length === 0) return;
-    uploadingImage = true;
-    uploadError = '';
-    try {
-      for (const file of Array.from(files)) {
-        const url = await uploadImage($ndk, file);
-        uploadedImages = [...uploadedImages, url];
-      }
-    } catch (err: any) {
-      uploadError = err?.message || 'Failed to upload image.';
-    } finally {
-      uploadingImage = false;
-      if (imageInputEl) imageInputEl.value = '';
-    }
-  }
-
-  async function handleVideoUpload(e: Event) {
-    const target = e.target as HTMLInputElement;
-    const files = target.files;
-    if (!files || files.length === 0) return;
-    uploadingVideo = true;
-    uploadError = '';
-    try {
-      for (const file of Array.from(files)) {
-        const url = await uploadVideo($ndk, file);
-        uploadedVideos = [...uploadedVideos, url];
-      }
-    } catch (err: any) {
-      uploadError = err?.message || 'Failed to upload video.';
-    } finally {
-      uploadingVideo = false;
-      if (videoInputEl) videoInputEl.value = '';
-    }
-  }
-
-  function removeImage(index: number) {
-    uploadedImages = uploadedImages.filter((_, i) => i !== index);
-  }
-
-  function removeVideo(index: number) {
-    uploadedVideos = uploadedVideos.filter((_, i) => i !== index);
-  }
-
-  // Post reply
-  async function postReply() {
-    if ((!replyText.trim() && uploadedImages.length === 0 && uploadedVideos.length === 0 && !pollConfig) || postingReply) return;
-
-    postingReply = true;
-    try {
-      if (replyComposerEl) {
-        replyText = mentionCtrl.extractText();
-        lastRenderedReply = replyText;
-      }
-
-      const ev = new NDKEvent($ndk);
-
-      // Check if this is a reply to a recipe comment (kind 1111)
-      // If the parent comment is kind 1111, use kind 1111 for nested reply
-      const isRecipeReply = event.kind === 1111;
-      ev.kind = pollConfig ? 1068 : (isRecipeReply ? 1111 : 1);
-      let replyContent = mentionCtrl.replacePlainMentions(replyText.trim());
-      const mediaUrls = [...uploadedImages, ...uploadedVideos];
-      if (mediaUrls.length > 0) {
-        const mediaText = mediaUrls.join('\n');
-        replyContent = replyContent ? `${replyContent}\n\n${mediaText}` : mediaText;
-      }
-      ev.content = replyContent;
-
-      // Reconstruct a minimal event object for the parent comment
-      const parentEventObj = {
-        id: event.id,
-        pubkey: event.pubkey,
-        kind: event.kind,
-        tags: event.tags
-      };
-
-      // For recipe replies, we need to get the root event info from the parent's tags
-      if (isRecipeReply) {
-        // Get root event info from parent comment's NIP-22 tags
-        const rootATag = event.getMatchingTags('A')[0] || event.getMatchingTags('a')[0];
-
-        if (rootATag) {
-          // Parse the address tag to extract root event info
-          const [kind, pubkey, ...dTagParts] = rootATag[1].split(':');
-          const dTag = dTagParts.join(':');
-          // Find the root event's actual ID from the parent's e tags (look for the
-          // e tag that references the root, not another comment)
-          const rootETag = event.getMatchingTags('E')[0] ||
-            event.getMatchingTags('e').find((t) => t[1] && t[1] !== event.id);
-          const rootEventObj = {
-            kind: parseInt(kind),
-            pubkey: pubkey,
-            id: rootETag?.[1] || '',
-            tags: [
-              ['d', dTag],
-              ['relay', rootATag[2] || '']
-            ]
-          };
-
-          // Use the utility with both root and parent event
-          ev.tags = buildNip22CommentTags(rootEventObj, parentEventObj);
-        } else {
-          // Fallback: parent lacks NIP-22 structure — build tags treating
-          // the parent as both root and parent (best effort)
-          ev.tags = buildNip22CommentTags(
-            {
-              ...parentEventObj,
-              kind: parentEventObj.kind ?? 1,
-              tags: parentEventObj.tags as string[][]
-            },
-            {
-              ...parentEventObj,
-              tags: parentEventObj.tags as string[][]
-            }
-          );
-        }
-      } else {
-        // For non-recipe replies, construct tags for standard note replies
-        // Try to derive the root event's pubkey from the parent event's tags.
-        // Prefer any 'p'/'P' tag, which typically references the root author.
-        const rootPubkeyFromTags = event.tags.find(
-          (t) => (t[0] === 'p' || t[0] === 'P') && t[1]
-        )?.[1];
-
-        const rootEventObj = {
-          kind: 1,
-          pubkey: rootPubkeyFromTags || event.pubkey,
-          id: mainEventId,
-          tags: []
-        };
-        ev.tags = buildNip22CommentTags(rootEventObj, parentEventObj);
-      }
-
-      const mentions = mentionCtrl.parseMentions(replyContent);
-      for (const pubkey of mentions.values()) {
-        if (!ev.tags.some((t) => t[0] === 'p' && t[1] === pubkey)) {
-          ev.tags.push(['p', pubkey]);
-        }
-      }
-
-      addClientTagToEvent(ev);
-      if (pollConfig) {
-        ev.tags.push(...buildPollTags(pollConfig));
-      }
-      await ev.publish();
-      replyText = '';
-      lastRenderedReply = '';
-      uploadedImages = [];
-      uploadedVideos = [];
-      uploadError = '';
-      pollConfig = null;
-      if (replyComposerEl) {
-        replyComposerEl.innerHTML = '';
-      }
-      mentionCtrl.resetMentionState();
-      showReplyBox = false;
-      refresh();
-    } catch {
-      // Failed to post reply
-    } finally {
-      postingReply = false;
-    }
+  function handleReplyPosted() {
+    showReplyBox = false;
+    refresh();
   }
 
   // Open zap modal
@@ -513,12 +302,7 @@
           <!-- Reply Button -->
           {#if $userPublickey}
             <button
-              on:click={() => {
-                showReplyBox = !showReplyBox;
-                if (showReplyBox && $userPublickey) {
-                  mentionCtrl.preloadFollowList();
-                }
-              }}
+              on:click={() => (showReplyBox = !showReplyBox)}
               class="action-btn action-btn-text"
             >
               {showReplyBox ? 'Cancel' : 'Reply'}
@@ -526,127 +310,19 @@
           {/if}
         </div>
 
-        <!-- Inline Reply Box -->
+        <!-- Inline Reply Composer -->
         {#if showReplyBox}
-          <div class="reply-form">
-            <div class="relative">
-              <div
-                bind:this={replyComposerEl}
-                class="reply-input"
-                contenteditable={!postingReply}
-                role="textbox"
-                tabindex="0"
-                aria-multiline="true"
-                data-placeholder="Add a reply..."
-                on:input={() => mentionCtrl.handleInput()}
-                on:keydown={(e) => mentionCtrl.handleKeydown(e)}
-                on:beforeinput={(e) => mentionCtrl.handleBeforeInput(e)}
-                on:paste={(e) => mentionCtrl.handlePaste(e)}
-              ></div>
-
-              <MentionDropdown
-                show={mentionState.showMentionSuggestions}
-                suggestions={mentionState.mentionSuggestions}
-                selectedIndex={mentionState.selectedMentionIndex}
-                searching={mentionState.mentionSearching}
-                query={mentionState.mentionQuery}
-                on:select={(e) => mentionCtrl.insertMention(e.detail)}
-              />
-            </div>
-            {#if uploadError}
-              <p class="text-red-500 text-xs">{uploadError}</p>
-            {/if}
-
-            {#if uploadedImages.length > 0}
-              <div class="flex flex-wrap gap-2">
-                {#each uploadedImages as imageUrl, index}
-                  <div class="relative group">
-                    <img src={imageUrl} alt="Upload preview" class="w-16 h-16 object-cover rounded-lg" style="border: 1px solid var(--color-input-border)" />
-                    <button type="button" on:click={() => removeImage(index)} class="absolute -top-2 -right-2 bg-red-500 hover:bg-red-600 text-white rounded-full p-0.5 shadow-lg" aria-label="Remove image">
-                      <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" /></svg>
-                    </button>
-                  </div>
-                {/each}
-              </div>
-            {/if}
-
-            {#if uploadedVideos.length > 0}
-              <div class="flex flex-wrap gap-2">
-                {#each uploadedVideos as videoUrl, index}
-                  <div class="relative group">
-                    <video src={videoUrl} class="w-24 h-16 object-cover rounded-lg" style="border: 1px solid var(--color-input-border)" preload="metadata" muted />
-                    <button type="button" on:click={() => removeVideo(index)} class="absolute -top-2 -right-2 bg-red-500 hover:bg-red-600 text-white rounded-full p-0.5 shadow-lg" aria-label="Remove video">
-                      <svg class="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" /></svg>
-                    </button>
-                  </div>
-                {/each}
-              </div>
-            {/if}
-
-            <div class="reply-buttons">
-              <label class="btn-media" class:opacity-50={uploadingImage || uploadingVideo || postingReply} title="Upload image">
-                <ImageIcon size={16} />
-                <input bind:this={imageInputEl} type="file" accept="image/*" class="sr-only" on:change={handleImageUpload} disabled={postingReply || uploadingImage || uploadingVideo} />
-              </label>
-              <label class="btn-media" class:opacity-50={uploadingImage || uploadingVideo || postingReply} title="Upload video">
-                <VideoIcon size={16} />
-                <input bind:this={videoInputEl} type="file" accept="video/*" class="sr-only" on:change={handleVideoUpload} disabled={postingReply || uploadingImage || uploadingVideo} />
-              </label>
-              <button
-                on:click={() => (showGifPicker = true)}
-                class="btn-gif"
-                title="Add GIF"
-                disabled={postingReply || uploadingImage || uploadingVideo}
-              >
-                <GifIcon size={16} />
-              </button>
-              <button
-                on:click={() => (showPollCreator = true)}
-                class="btn-gif"
-                title="Create poll"
-                disabled={postingReply}
-                class:opacity-50={postingReply}
-              >
-                <ChartBarHorizontalIcon size={16} class={pollConfig ? 'text-primary' : ''} />
-              </button>
-              {#if uploadingImage}
-                <span class="text-xs text-caption">Uploading image...</span>
-              {:else if uploadingVideo}
-                <span class="text-xs text-caption">Uploading video...</span>
-              {/if}
-              {#if pollConfig}
-                <span class="text-xs text-orange-600 flex items-center gap-1">
-                  <ChartBarHorizontalIcon size={12} />
-                  Poll ({pollConfig.options.length})
-                  <button type="button" on:click={() => (pollConfig = null)} class="hover:text-orange-800">×</button>
-                </span>
-              {/if}
-              <button
-                on:click={postReply}
-                disabled={(!replyText.trim() && uploadedImages.length === 0 && uploadedVideos.length === 0 && !pollConfig) || postingReply || uploadingImage || uploadingVideo}
-                class="btn-post"
-              >
-                {postingReply ? 'Posting...' : 'Post'}
-              </button>
-              <button
-                on:click={() => {
-                  showReplyBox = false;
-                  replyText = '';
-                  lastRenderedReply = '';
-                  uploadedImages = [];
-                  uploadedVideos = [];
-                  uploadError = '';
-                  if (replyComposerEl) {
-                    replyComposerEl.innerHTML = '';
-                  }
-                  mentionCtrl.resetMentionState();
-                }}
-                class="btn-cancel"
-              >
-                Cancel
-              </button>
-            </div>
-          </div>
+          <ReplyComposer
+            parentEvent={rootEvent}
+            replyTo={event}
+            placeholder="Add a reply..."
+            signingStrategy="implicit"
+            onErrorStrategy="silent"
+            compact
+            showCancel
+            onPosted={handleReplyPosted}
+            on:cancel={() => (showReplyBox = false)}
+          />
         {/if}
       </div>
     </div>
@@ -657,20 +333,6 @@
     <ZapModal bind:open={zapModalOpen} {event} />
   {/if}
 {/if}
-
-<GifPicker
-  bind:open={showGifPicker}
-  on:select={(e) => {
-    uploadedImages = [...uploadedImages, e.detail.url];
-  }}
-/>
-
-<PollCreator
-  bind:open={showPollCreator}
-  on:create={(e) => {
-    pollConfig = e.detail;
-  }}
-/>
 
 <style>
   /* Comment card - full width, no nesting */
@@ -804,90 +466,5 @@
     align-items: center;
     gap: 0.25rem;
     color: var(--color-text-secondary);
-  }
-
-  /* Reply form */
-  .reply-form {
-    margin-top: 0.5rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-
-  .reply-input {
-    width: 100%;
-    padding: 0.375rem 0.5rem;
-    font-size: 0.875rem;
-    border-radius: 0.5rem;
-    background: var(--color-input-bg);
-    border: 1px solid var(--color-input-border);
-    color: var(--color-text-primary);
-    white-space: pre-wrap;
-    word-break: break-word;
-  }
-
-  .reply-input:focus {
-    outline: none;
-    ring: 2px;
-    ring-color: var(--color-primary);
-  }
-
-  .reply-input:empty:before {
-    content: attr(data-placeholder);
-    color: var(--color-caption);
-    pointer-events: none;
-  }
-
-  .reply-buttons {
-    display: flex;
-    gap: 0.5rem;
-  }
-
-  .btn-post {
-    padding: 0.375rem 0.75rem;
-    font-size: 0.75rem;
-    font-weight: 500;
-    color: white;
-    background: var(--color-primary);
-    border-radius: 0.5rem;
-  }
-
-  .btn-post:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-  }
-
-  .btn-cancel {
-    padding: 0.375rem 0.75rem;
-    font-size: 0.75rem;
-    font-weight: 500;
-    color: var(--color-text-primary);
-    background: var(--color-input);
-    border-radius: 0.5rem;
-  }
-
-  .btn-cancel:hover {
-    opacity: 0.8;
-  }
-
-  .btn-gif,
-  .btn-media {
-    padding: 0.375rem;
-    color: var(--color-caption);
-    border-radius: 0.375rem;
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-    transition: opacity 0.15s;
-  }
-
-  .btn-gif:hover,
-  .btn-media:hover {
-    opacity: 0.7;
-  }
-
-  .btn-gif:disabled {
-    opacity: 0.4;
-    cursor: not-allowed;
   }
 </style>
